DTO fixes

+++ src/main/java/io/contek/invoker/binancelinear/api/common/_BookTicker.java
@@ -1,14 +1,15 @@
 package io.contek.invoker.binancelinear.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
+import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _BookTicker {
 
   public String symbol;
-  public Double bidPrice;
-  public Double bidQty;
-  public Double askPrice;
-  public Double askQty;
-  public Long time;
+  public BigDecimal bidPrice;
+  public BigDecimal bidQty;
+  public BigDecimal askPrice;
+  public BigDecimal askQty;
+  public long time;
 }
